refactor: integrate LUT calibration functions into module `hrl.luts`
